fix(encode): use a `for`-loop, instead of `next()`
For `next()`, the order in which the indices are enumerated is not
specified[1], even for numeric indices. Even though it seemed to work,
now the encode-/decode-logic uses a numerical `for`-loop, which should
keep the order and not miss any items after a `nil`-item.

Also, the loop inside the `encode`-function, uses the `list._length`
instead of `#items`; which seems to ignore data, when there're 2 or more
`nil`-values!?

[1] https://www.lua.org/manual/5.1/manual.html#pdf-next